feat upload image method

+++ src/main/java/com/betacom/services/interfaces/InterfaceUploadService.java
@@ -3,7 +3,7 @@ package com.betacom.services.interfaces;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface InterfaceUploadService {
-	String saveImage(MultipartFile file, Integer id) throws Exception;
+	String saveImage(MultipartFile file, Long id) throws Exception;
 	
 	void removeImage(String fileName) throws Exception;
